+++ rpgsaga/saga/src/doublyLinkedList/doublyLinkedListNode.ts
@@ -1,4 +1,4 @@
-import { StatusEffect } from '../statusEffect/statusEffect';
+import { StatusEffect } from '../spell_system/statusEffect/statusEffect';
 
 export class DoublyLinkedListNode {
   public value: StatusEffect;
