+++ rpgsaga/saga/src/spell_system/spell/knightAttack.ts
@@ -1,4 +1,4 @@
-import { Message } from '../message';
+import { Message } from '../../message';
 
 import { Spell } from './spell';
