+++ rpgsaga/saga/src/spell_system/statusEffect/statusEffect.ts
@@ -1,4 +1,4 @@
-import { Character } from '../character';
+import { Character } from '../../character';
 
 export abstract class StatusEffect {
   private _name: string;
