+++ rpgsaga/saga/src/spell_system/statusEffect/stunEffect.ts
@@ -1,4 +1,4 @@
-import { Character } from '../character';
+import { Character } from '../../character';
 
 import { StatusEffect } from './statusEffect';
